Review Detail UI 작업 및 모델링

+++ app/src/main/java/com/univreview/model/AbstractDataProvider.java
@@ -4,6 +4,6 @@ package com.univreview.model;
  * Created by DavidHa on 2017. 1. 24..
  */
 public abstract class AbstractDataProvider {
-    public abstract int getId();
+    public abstract long getId();
     public abstract String getName();
 }
